+++ src/rules/custom-property-pattern/__tests__/index.js
@@ -15,6 +15,16 @@ testRule(/foo-.+/, tr => {
   tr.notOk(":root { --foo-: 0; }", messages.expected)
 })
 
+// Same as above but with a string instead of a RegExp
+testRule("foo-.+", tr => {
+  warningFreeBasics(tr)
+
+  tr.ok(":root { --foo-bar: 0; }")
+  tr.ok(":root { --boo-foo-bar: 0; }")
+  tr.notOk(":root { --boo-bar: 0; }", messages.expected)
+  tr.notOk(":root { --foo-: 0; }", messages.expected)
+})
+
 testRule(/^[A-Z][a-z]+-[a-z][a-zA-Z]+$/, tr => {
   warningFreeBasics(tr)